urbanpulse-ai/requirements.txt
fastapi
uvicorn
requests
python-multipart
pillow
opencv-python
numpy
pandas
scikit-learn
joblib
python-dotenv